requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
python-dotenv==1.1.1
pillow
celery==5.5.3
gunicorn==23.0.0
redis>=4.6.0
kombu>=5.3
# Testing
requests
pytest
django 
pytest-django
